backend/requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
django-filter>=24.0,<25.0
django-environ>=0.11,<1.0
django-storages>=1.14,<2.0
django-import-export>=4.0,<5.0
django-redis>=5.4,<6.0
django-ratelimit>=4.1,<5.0
drf-spectacular>=0.27,<1.0
celery>=5.4,<6.0
django-celery-beat>=2.6,<3.0
redis>=5.0,<6.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
Pillow>=10.0,<11.0
psycopg2-binary>=2.9,<3.0
gunicorn>=22.0,<23.0
uvicorn[standard]>=0.30,<1.0
WeasyPrint>=62.0,<63.0
whitenoise>=6.7,<7.0
paynow>=1.0,<2.0
stripe>=10.0,<11.0
openpyxl>=3.1,<4.0
